+++ src/TestExecution/DebugSessionManager.ts
@@ -1,15 +1,17 @@
 import { inject, injectable } from 'inversify';
 import type { WorkspaceFolder } from 'vscode';
 import { debug } from 'vscode';
-import { Mode, type Xdebug } from '../PHPUnit';
+import type { ProcessBuilder } from '../PHPUnit';
 import { TYPES } from '../types';
 
 @injectable()
 export class DebugSessionManager {
     constructor(@inject(TYPES.WorkspaceFolder) private workspaceFolder: WorkspaceFolder) {}
 
-    async wrap(xdebug: Xdebug | undefined, fn: () => Promise<void>): Promise<void> {
-        if (xdebug?.mode === Mode.debug) {
+    async wrap(builder: ProcessBuilder, fn: () => Promise<void>): Promise<void> {
+        const xdebug = builder.getXdebug();
+
+        if (xdebug?.isDebugMode()) {
             // TODO(#346): await debug session attachment before running tests
             await debug.startDebugging(
                 this.workspaceFolder,
@@ -20,7 +22,7 @@ export class DebugSessionManager {
         try {
             await fn();
         } finally {
-            if (xdebug?.mode === Mode.debug && debug.activeDebugSession?.type === 'php') {
+            if (xdebug?.isDebugMode() && debug.activeDebugSession?.type === 'php') {
                 debug.stopDebugging(debug.activeDebugSession);
             }
         }