+++ src/TestExecution/ProcessBuilderFactory.ts
@@ -1,7 +1,10 @@
 import { inject, injectable } from 'inversify';
 import { TestRunProfileKind } from 'vscode';
 import { Configuration } from '../Configuration';
-import { Mode, PHPUnitXML, ProcessBuilder, Xdebug } from '../PHPUnit';
+import { type Path, PathReplacer } from '../PHPUnit/PathReplacer';
+import { PHPUnitXML } from '../PHPUnit/PHPUnitXML';
+import { ProcessBuilder } from '../PHPUnit/ProcessBuilder/ProcessBuilder';
+import { Mode, Xdebug } from '../PHPUnit/ProcessBuilder/Xdebug';
 
 @injectable()
 export class ProcessBuilderFactory {
@@ -11,11 +14,10 @@ export class ProcessBuilderFactory {
     ) {}
 
     async create(profileKind?: TestRunProfileKind): Promise<ProcessBuilder> {
-        const builder = new ProcessBuilder(this.config, { cwd: this.phpUnitXML.root() });
-        const xdebug = new Xdebug(this.config);
-        builder.setXdebug(xdebug);
-        await xdebug.setMode(this.toMode(profileKind));
-        return builder;
+        const options = { cwd: this.phpUnitXML.root() };
+        const pathReplacer = new PathReplacer(options, this.config.get('paths') as Path);
+        const xdebug = await new Xdebug(this.config).setMode(this.toMode(profileKind));
+        return new ProcessBuilder(this.config, options, pathReplacer, xdebug);
     }
 
     private toMode(profileKind?: TestRunProfileKind): Mode | undefined {